+++ src/engines/hf/hf_es_engine_class.F90
@@ -17,48 +17,55 @@
 !  along with this program. If not, see <https://www.gnu.org/licenses/>.
 !
 !
-module hf_geoopt_engine_class
+module hf_es_engine_class
 !
 !!
-!! HF geometry optimization engine class
-!! Written by Sarai D. Folkestad, Eirik F. Kjønstad, and Alexander C. Paul, 2018-2022
+!! HF excited state engine class
+!! Written by Sarai D. Folkestad, Alexander C. Paul, 2018-2022
 !!
 !
    use hf_engine_class, only: hf_engine
    use hf_class,        only: hf
 !
-   use hf_geoopt_task_class, only: hf_geoopt_task
+   use tdhf_task_class, only: tdhf_task
 !
    implicit none
 !
-   type, extends(hf_engine) :: hf_geoopt_engine
+   type, extends(hf_engine) :: hf_es_engine
 !
-      type(hf_geoopt_task), allocatable, private :: geometry_optimization
+      type(tdhf_task), allocatable, private :: tdhf
 !
    contains
 !
-      procedure, public :: ignite => ignite_hf_geoopt_engine
+      procedure, public :: ignite => ignite_hf_es_engine
 !
-   end type hf_geoopt_engine
+   end type hf_es_engine
 !
 contains
 !
 !
-   subroutine ignite_hf_geoopt_engine(this, wf)
+   subroutine ignite_hf_es_engine(this, wf)
 !!
 !!    Ignite
 !!    Written by Alexander C. Paul, May 2022
 !!
+      use hf_gs_engine_class, only: hf_gs_engine
+!
       implicit none
 !
-      class(hf_geoopt_engine), intent(inout) :: this
-      class(hf), intent(inout) :: wf
+      class(hf_es_engine), intent(inout) :: this
+      class(hf),           intent(inout) :: wf
 !
-      this%geometry_optimization = hf_geoopt_task()
-      call this%geometry_optimization%execute(wf)
+      type(hf_gs_engine), allocatable :: gs_engine
 !
-   end subroutine ignite_hf_geoopt_engine
+      gs_engine = hf_gs_engine()
+      call gs_engine%ignite(wf)
+!
+      this%tdhf = tdhf_task()
+      call this%tdhf%execute(wf)
+!
+   end subroutine ignite_hf_es_engine
 !
 !
-   end module hf_geoopt_engine_class
+   end module hf_es_engine_class